ft: add warehouses component
